# Optimize UTF-8 encoding by avoiding String.getBytes allocations