@richardstartin 2022-09-13
Which JDK version are these numbers for? I _seem_ to recall this only making sense on JDK8 when I wrote the code originally, but if these numbers are actually JDK11 then that's great! If not, please post JDK11+ numbers too.