@bantonsson 2022-09-13
I'm wondering why we end up there, and there is no _EncodingCache_ used in the writer 🤷🏼‍♂️