@richardstartin 2022-09-13
> I'm wondering why we end up there, and there is no _EncodingCache_ used in the writer 🤷🏼‍♂️

I removed the encoding cache a long time ago because it was counterproductive and unbounded and `String.getBytes` appeared to be working so well in JDK11 it wasn't worth the bother.