@richardstartin 2022-09-13
> Ah, now I remember. It shouldn't need to be unbounded though, and the same machinery introduced here could live inside the _cache_. Would make it easier to switch implementations based on the JDK as well.

The encoding cache was a mistake, `UTF8BytesString` for all the cacheable or constant strings which can be shared across spans is much better.